Add q_sort
Implement sort by using merge sort.

The first step of merge sort is to divide the queue, so I used
list_cut_position() to divide to queue into two, head and head2.
Then I recursively call the function to divide it into smaller
queues. I merge the queues by comparing the string with strcmp()
and insert the smaller nodes to the tail of head.